test: fix naming warnings in tests
- Rename test functions to snake_case:
  - test_abstractPropertyNegative_errors -> test_abstract_property_negative_errors
  - test_async_stream_pattern_getReader -> test_async_stream_pattern_get_reader
  - test_no_predicate_NOT_compatible_with_type_guard -> test_no_predicate_not_compatible_with_type_guard
  - test_type_guard_NOT_compatible_with_asserts -> test_type_guard_not_compatible_with_asserts

- Rename test variables to snake_case (avoid conflicts):
  - lit_A, lit_B -> lit_a_upper, lit_b_upper
  - lit_ABC, lit_DEF -> lit_abc_upper, lit_def_upper
  - lit_Name, lit_Value -> lit_name_upper, lit_value_upper

- Restore type imports in solver test files that were removed by auto-fix

Final status:
- 0 clippy warnings (lib)
- 10 clippy warnings (tests, excluding "file is loaded as a module multiple times")
- 0 cargo build warnings
- All 10,110 tests passing

+++ src/lsp/folding.rs
@@ -328,10 +328,10 @@ impl<'a> FoldingRangeProvider<'a> {
         // Get line positions
         let start_pos = self
             .line_map
-            .offset_to_position(lo as u32, self.source_text);
+            .offset_to_position(lo, self.source_text);
         let end_pos = self
             .line_map
-            .offset_to_position(hi.saturating_sub(1) as u32, self.source_text);
+            .offset_to_position(hi.saturating_sub(1), self.source_text);
 
         (start_pos.line, end_pos.line)
     }
